Extracted balance validating fn
